shimacos/pyproject.toml
[tool.poetry]
name = "kaggle-otto-3rd-place-solution"
version = "0.1.0"
description = ""
authors = ["Naoto Shimakoshi"]

[tool.poetry.dependencies]
python = "^3.8,<3.9"
scipy = "^1.7.1"
scikit-learn = "^1.1.1"
tqdm = "^4.62.1"
Jinja2 = "^3.0.1"
joblib = "1.0.0"
retry = "^0.9.2"
jupyter = "^1.0.0"
matplotlib = "^3.5.1"
invoke = "^1.7.0"
hydra-core = "^1.1.1"
seaborn = "^0.11.2"
japanize-matplotlib = "^1.1.3"
timeout-decorator = "^0.5.0"
lightgbm = "^3.3.2"
kfp = "^1.8.12"
setuptools = "^62.4.0"
pandas = "^1.4.4"
pandas-gbq = "^0.17.8"
google-cloud-bigquery = "2.6.1"
google-cloud-bigquery-storage = "^2.15.0"
google-cloud-aiplatform = "^1.17.0"
gcsfs = "^2022.8.2"
fsspec = "^2022.8.2"
timm = "^0.6.11"
xgboost = "^1.6.2"
catboost = "^1.1"
hydra-colorlog = "^1.2.0"
wandb = "^0.13.4"
igraph = "^0.10.1"
networkx = "^2.8.7"
pyvis = "^0.3.0"
gensim = "^4.3.0"
transformers = "^4.25.1"
opencv-python = "^4.7.0.68"
torch = "^1.13.1"
implicit = "^0.6.2"
polars = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "22.3.0"
autopep8 = "^1.5.7"
flake8 = "^3.9.2"
ipdb = "^0.13.9"
functions-framework = "^3.1.0"
pre-commit = "^2.20.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py36', 'py37']

[tool.sqlfluff.core]
verbose = 1
exclude_rules = 'L032'

[tool.sqlfluff.templater.jinja.context]

start_date = '2022-12-01'
end_date = '2022-12-14'
project_id = 'hoge'
dataset_id = 'hoge'
dataset_name = 'hoge'
sql_name = 'hoge'
split_ts = "2022-08-21T22:00:00"
start_ts = "2022-08-14T22:00:00"
end_ts = "2022-09-04T22:00:00"
minutes = [1, 15, 30]
hours = [1, 3, 5, 12, 24, 48, 72]
days = [1, 3, 5, 7, 14]

[tool.sqlfluff.rules]
tab_space_size = 2

[tool.sqlfluff.rules.L010]
capitalisation_policy = 'upper'

[tool.sqlfluff.rules.L014]
extended_capitalisation_policy = 'lower'
unquoted_identifiers_policy = 'aliases'

[tool.sqlfluff.rules.L016]
max_line_length = 200

[tool.sqlfluff.rules.L064]
preferred_quoted_literal_style = 'double_quotes'
